requirements.txt
streamlit
pandas
numpy
openpyxl
PyInstaller
altair
python-dateutil
vl-convert-python
weasyprint
scipy
matplotlib
pyarrow
pydeck
watchdog